+++ test/unit_test/passes/pytorch/test_autogptq.py
@@ -11,7 +11,7 @@ from olive.data.config import DataComponentConfig, DataConfig
 from olive.hardware.accelerator import AcceleratorSpec, Device
 from olive.model import HfModelHandler
 from olive.passes.olive_pass import create_pass_from_dict
-from olive.passes.pytorch.gptq import GptqQuantizer
+from olive.passes.pytorch.autogptq import GptqQuantizer
 from test.unit_test.utils import make_local_tiny_llama
 
 test_gptq_dc_config = DataConfig(
